+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.23.4-preview] - 2021-12-07
+- Fixed platform-dependent line-endings in output of UpdateConfigurationValidation.
+- Fixed ApiValidation flagging new overrides as additions.
+- Changed API Updater Configuration Validation to ignore assemblies that are not part of the final package.
+- Introduced a way for devs to bypass ApiValidation sanity checks to be able to run tests locally regardless of OS/Unity version.
+- Fixed missing assembly qualification in ApiUpdater Configuration Validator tests.
+
 ## [0.23.3-preview] - 2021-11-10
 - Fixed `Mono.Cecil.AssemblyResolutionException` while running API Updater Configuration Validation failing to resolve assemblies under `Library/ScriptAssemblies` folder.